refactor(json-crdt): 💡 move proxy types a folder lower
